Cleaner implementation of leading foot methods
Now, all methods are called from the main method identifyLeadingFoot(). The implementation of this is slightly different for Gait vs the GaitCycle class - this is because the GaitCycle can leverage implicit knowledge to make use of marker/IK data; the Gait class cannot.

There remains an additional function identifyFrontFootMarkers, which is needed in the current implementation of the extrapolated boundary of support stuff.